+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [Unreleased][unreleased]
+
 - Fix lost json subfields checking
 
 ## [1.2.3][] - 2021-05-22